apps/shop/models.py
from turtle import mode
from autoslug import AutoSlugField
from django.db import models
from apps.common.models import BaseModel, IsDeletedModel
from apps.sellers.models import Seller
from apps.accounts.models import User


class Category(BaseModel):

    name = models.CharField(
        max_length=100, unique=True
    )
    slug = AutoSlugField(
        populate_from="name", unique=True, always_update=True
    )
    image = models.ImageField(upload_to='category_images/')

    def __str__(self):
        return str(self.name)

    class Meta:
        verbose_name_plural = "Categories"


class Product(IsDeletedModel):

    seller = models.ForeignKey(
        Seller, on_delete=models.SET_NULL,
        related_name="products",
        null=True
    )
    name = models.CharField(max_length=100)
    slug = AutoSlugField(
        populate_from="name", unique=True, db_index=True
    )
    desc = models.TextField()
    price_old = models.DecimalField(
        max_digits=10, decimal_places=2, null=True
    )
    price_current = models.DecimalField(
        max_digits=10, decimal_places=2
    )
    category = models.ForeignKey(
        Category, on_delete=models.CASCADE,
        related_name="products"
    )
    in_stock = models.IntegerField(default=5)

    image1 = models.ImageField(upload_to="product_images/")
    image2 = models.ImageField(upload_to="product_images/", blank=True)
    image3 = models.ImageField(upload_to="product_images/", blank=True)

    def __str__(self):
        return str(self.name)


RATING_CHOICES = [
    (1, 1), (2, 2), (3, 3), (4, 4), (5, 5)
]


class Review(IsDeletedModel):
    """
    Модель данных для отзывов о продукте (Product)
    """
    user = models.ForeignKey(
        # каждый отзыв связан с одним пользователем
        User, on_delete=models.CASCADE, related_name="reviews"
    )
    product = models.ForeignKey(
        # каждый отзыв связан с одним продуктом
        Product, on_delete=models.CASCADE, related_name="reviews"
    )
    rating = models.SmallIntegerField(
        choices=RATING_CHOICES, default=None, null=True, blank=True
    )
    text = models.TextField(
        null=True, blank=True
    )
