requirements.txt
asgiref==3.10.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-autoslug==1.9.9
django-filter==25.2
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
djangorestframework-stubs==3.16.5
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
pillow==12.0.0
PyJWT==2.10.1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
